Now serial driver API now knows that buffers are 256 bytes.

+++ back/serial.h
@@ -13,23 +13,23 @@ typedef enum serial_error_bit {
     SE_PARITY_ERROR = _BV(UPE0),  // 0x04
 } serial_error_bit;
 
-extern void     init_serial            (void);
+extern void    init_serial            (void);
 
-extern void     serial_rx_start        (void);
-extern uint8_t  serial_rx_errors       (void);
-extern uint8_t  serial_rx_peek_errors  (void);
-extern bool     serial_rx_has_chars    (void);
-extern bool     serial_rx_has_lines    (void);
-extern uint16_t serial_rx_char_count   (void);
-extern uint16_t serial_rx_line_count   (void);
-extern uint8_t  serial_rx_peek_char    (uint16_t pos);
-extern void     serial_rx_consume      (uint16_t count);
+extern void    serial_rx_start        (void);
+extern uint8_t serial_rx_errors       (void);
+extern uint8_t serial_rx_peek_errors  (void);
+extern bool    serial_rx_has_chars    (void);
+extern bool    serial_rx_has_lines    (void);
+extern uint8_t serial_rx_char_count   (void);
+extern uint8_t serial_rx_line_count   (void);
+extern uint8_t serial_rx_peek_char    (uint8_t pos);
+extern void    serial_rx_consume      (uint8_t count);
 
-extern uint8_t  serial_tx_errors       (void);
-extern uint8_t  serial_tx_peek_errors  (void);
-extern bool     serial_tx_is_idle      (void);
-extern uint8_t  serial_tx_is_available (void);
-extern uint8_t  serial_tx_char_count   (void);
-extern bool     serial_tx_put_char     (uint8_t c);
+extern uint8_t serial_tx_errors       (void);
+extern uint8_t serial_tx_peek_errors  (void);
+extern bool    serial_tx_is_idle      (void);
+extern uint8_t serial_tx_is_available (void);
+extern uint8_t serial_tx_char_count   (void);
+extern bool    serial_tx_put_char     (uint8_t c);
 
 #endif /* !SERIAL_included */
